Move TrueDot out of sandbox.
